scripts/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0,<2.6
torchaudio>=2.0,<2.6
huggingface_hub>=0.20,<1.0
faster-whisper>=1.0.0
pyannote.audio>=3.1,<4.0
